+++ CHANGELOG.md
@@ -5,10 +5,19 @@
 ### Changed
 
 * ClientInterface supports PSR7 requests and responses. 
+* Removed type hint for second parameter for `AbstractCurl::prepare`.
+* Removed type hint for second parameter for `RequestException::setRequest`.
+* `RequestException` supports both PSR-7 requests and Buzz requests. 
 
 ### Added 
 
 * Added Request and Response converters
+* Added `Curl::sendRequest()`, `MultiCurl::sendRequest()` and `FileGetContents::sendRequest()` that
+  supports sending PSR-7 requests.   
+
+### Deprecated
+
+*`Curl::send()`, `MultiCurl::send()` and `FileGetContents::send()` in favor for `sendRequest()`. 
 
 ## 0.15.2